Rename publicProfile into PublicProfilePage and PublicProfile component into index
