+++ client/src/App.tsx
@@ -9,7 +9,7 @@ import ProfileEditor from './components/ProfileEditor';
 import UpdateEmail from './components/UpdateEmail';
 import AlertProvider from './components/AlertProvider';
 import AlertSnackBar from './components/AlertSnackBar';
-import PublicProfile from './components/PublicProfile';
+import PublicProfilePage from './components/PublicProfile/PublicProfilePage';
 import VisitHistory from './components/VisitHistory';
 import Matches from './components/Matches';
 import Blocks from './components/Blocks';
@@ -107,7 +107,7 @@ const App = () => {
 										}
 									/>
 									<Route path="/profile" element={<ProfileEditor />} />
-									<Route path="/profile/:id" element={<PublicProfile />} />
+									<Route path="/profile/:id" element={<PublicProfilePage />} />
 									<Route path="/update_email" element={<UpdateEmail />} />
 									<Route path="/visit_history" element={<VisitHistory />} />
 									<Route path="/likes" element={<Likes />} />
